MULE-20025: Move StreamingManager to the base registry
