+++ core/src/main/java/org/mule/runtime/core/api/streaming/DefaultStreamingManager.java
@@ -6,9 +6,6 @@
  */
 package org.mule.runtime.core.api.streaming;
 
-import static java.lang.Boolean.parseBoolean;
-import static java.lang.String.format;
-import static java.lang.System.getProperty;
 import static org.mule.runtime.api.i18n.I18nMessageFactory.createStaticMessage;
 import static org.mule.runtime.api.util.MuleSystemProperties.MULE_ENABLE_STREAMING_STATISTICS;
 import static org.mule.runtime.core.api.lifecycle.LifecycleUtils.disposeIfNeeded;
@@ -17,6 +14,11 @@ import static org.mule.runtime.core.api.util.ClassUtils.instantiateClass;
 import static org.mule.runtime.core.api.util.IOUtils.closeQuietly;
 import static org.mule.runtime.core.internal.util.CompositeClassLoader.from;
 import static org.mule.runtime.core.privileged.util.EventUtils.getRoot;
+
+import static java.lang.Boolean.parseBoolean;
+import static java.lang.String.format;
+import static java.lang.System.getProperty;
+
 import static org.slf4j.LoggerFactory.getLogger;
 
 import org.mule.api.annotation.NoExtend;
@@ -24,7 +26,6 @@ import org.mule.runtime.api.event.EventContext;
 import org.mule.runtime.api.lifecycle.Disposable;
 import org.mule.runtime.api.lifecycle.Initialisable;
 import org.mule.runtime.api.lifecycle.InitialisationException;
-import org.mule.runtime.api.scheduler.SchedulerService;
 import org.mule.runtime.api.streaming.Cursor;
 import org.mule.runtime.api.streaming.CursorProvider;
 import org.mule.runtime.api.util.Pair;
@@ -73,9 +74,6 @@ public class DefaultStreamingManager implements StreamingManager, Initialisable,
 
   @Inject
   private MuleContext muleContext;
-
-  @Inject
-  private SchedulerService schedulerService;
 
   @Inject
   private StreamingGhostBuster ghostBuster;
