+++ modules/spring-config/src/main/java/org/mule/runtime/config/internal/BaseSpringMuleContextServiceConfigurator.java
@@ -10,19 +10,27 @@ import static org.mule.runtime.api.config.FeatureFlaggingService.FEATURE_FLAGGIN
 import static org.mule.runtime.core.api.config.MuleProperties.MULE_PROFILING_SERVICE_KEY;
 import static org.mule.runtime.core.api.config.MuleProperties.OBJECT_SCHEDULER_BASE_CONFIG;
 import static org.mule.runtime.core.api.config.MuleProperties.OBJECT_SCHEDULER_POOLS_CONFIG;
+import static org.mule.runtime.core.api.config.MuleProperties.OBJECT_STREAMING_GHOST_BUSTER;
+import static org.mule.runtime.core.api.config.MuleProperties.OBJECT_STREAMING_MANAGER;
 import static org.mule.runtime.core.internal.exception.ErrorTypeLocatorFactory.createDefaultErrorTypeLocator;
 
 import org.mule.runtime.api.artifact.Registry;
 import org.mule.runtime.api.component.location.ConfigurationComponentLocator;
+import org.mule.runtime.api.config.custom.ServiceConfigurator;
 import org.mule.runtime.api.exception.ErrorTypeRepository;
 import org.mule.runtime.api.scheduler.SchedulerContainerPoolsConfig;
+import org.mule.runtime.api.service.Service;
 import org.mule.runtime.config.internal.context.BaseConfigurationComponentLocator;
+import org.mule.runtime.config.internal.factories.SchedulerBaseConfigFactory;
 import org.mule.runtime.core.api.MuleContext;
+import org.mule.runtime.core.api.registry.SpiServiceRegistry;
+import org.mule.runtime.core.api.streaming.DefaultStreamingManager;
 import org.mule.runtime.core.internal.config.CustomService;
 import org.mule.runtime.core.internal.config.CustomServiceRegistry;
 import org.mule.runtime.core.internal.exception.ContributedErrorTypeLocator;
 import org.mule.runtime.core.internal.exception.ContributedErrorTypeRepository;
 import org.mule.runtime.core.internal.profiling.NoOpProfilingService;
+import org.mule.runtime.core.internal.streaming.StreamingGhostBuster;
 import org.mule.runtime.core.privileged.exception.ErrorTypeLocator;
 
 import java.util.Map;
@@ -69,17 +77,28 @@ class BaseSpringMuleContextServiceConfigurator extends AbstractSpringMuleContext
     contributedErrorTypeLocator.setDelegate(createDefaultErrorTypeLocator(contributedErrorTypeRepository));
     registerConstantBeanDefinition(ErrorTypeLocator.class.getName(), contributedErrorTypeLocator);
 
+    loadServiceConfigurators();
+
+    registerBeanDefinition(OBJECT_STREAMING_MANAGER, getBeanDefinition(DefaultStreamingManager.class));
+    registerBeanDefinition(OBJECT_STREAMING_GHOST_BUSTER, getBeanDefinition(StreamingGhostBuster.class));
+
     registerBeanDefinition(OBJECT_SCHEDULER_POOLS_CONFIG,
                            getConstantObjectBeanDefinition(SchedulerContainerPoolsConfig.getInstance()));
     registerBeanDefinition(OBJECT_SCHEDULER_BASE_CONFIG, getBeanDefinition(SchedulerBaseConfigFactory.class));
 
     registerBeanDefinition(MULE_PROFILING_SERVICE_KEY, getBeanDefinition(NoOpProfilingService.class));
 
-    createRuntimeServices();
+    createCustomServices();
     absorbOriginalRegistry();
   }
 
-  private void createRuntimeServices() {
+  private void loadServiceConfigurators() {
+    new SpiServiceRegistry()
+        .lookupProviders(ServiceConfigurator.class, Service.class.getClassLoader())
+        .forEach(customizer -> customizer.configure(getCustomServiceRegistry()));
+  }
+
+  private void createCustomServices() {
     final Map<String, CustomService> customServices = getCustomServiceRegistry().getCustomServices();
     for (String serviceName : customServices.keySet()) {
 
@@ -88,12 +107,10 @@ class BaseSpringMuleContextServiceConfigurator extends AbstractSpringMuleContext
       }
 
       final CustomService customService = customServices.get(serviceName);
-      // TODO MULE-19927 get these form a more specific place and avoid this filter
-      if (isServiceRuntimeProvided(customService)) {
-        final BeanDefinition beanDefinition = getCustomServiceBeanDefinition(customService, serviceName);
 
-        registerBeanDefinition(serviceName, beanDefinition);
-      }
+      final BeanDefinition beanDefinition = getCustomServiceBeanDefinition(customService, serviceName);
+
+      registerBeanDefinition(serviceName, beanDefinition);
     }
   }
