+++ modules/spring-config/src/main/java/org/mule/runtime/config/internal/context/service/InjectParamsFromContextServiceMethodInvoker.java
@@ -4,7 +4,7 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.runtime.config.internal;
+package org.mule.runtime.config.internal.context.service;
 
 import static java.lang.String.format;
 import static java.lang.reflect.Modifier.isPublic;
