+++ modules/spring-config/src/main/java/org/mule/runtime/config/internal/factories/SchedulerBaseConfigFactory.java
@@ -4,7 +4,7 @@
  * license, a copy of which has been included with this distribution in the
  * LICENSE.txt file.
  */
-package org.mule.runtime.config.internal;
+package org.mule.runtime.config.internal.factories;
 
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
 import static org.mule.runtime.api.scheduler.SchedulerConfig.config;
